requirements-pinned.txt


# Pinned versions for production deployment

# Install with: pip install -r requirements-pinned.txt

# Core dependencies
pandas==1.5.3
numpy==1.23.5
requests==2.28.2
python-dotenv==0.21.0
pyyaml==6.0
supabase==1.0.1
psycopg2-binary==2.9.5
tenacity==8.1.0
pytz==2022.7
python-dateutil==2.8.2
click==8.1.3
apache-airflow==2.5.1
apache-airflow-providers-postgres==5.0.0
apache-airflow-providers-http==4.0.1

# API clients
alpha-vantage==2.3.1
finnhub-python==1.3.2
fredapi==0.5.1
pycoingecko==3.1.1

# Data processing
sqlalchemy==2.0.0
alembic==1.9.1
marshmallow==3.19.0
APScheduler

# Monitoring & logging
python-json-logger==2.0.7
sentry-sdk==1.15.0

# Testing (included for CI/CD)
pytest==7.2.2
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-asyncio==0.20.3

# Code quality (included for CI/CD)
black==23.1.0
flake8==6.0.0
isort==5.12.0
mypy==1.0.0
pylint==2.16.4
bandit==1.7.5
safety==2.3.5
